refactor game.js: create service (answerService, questionService, gameService, roundService, randomService)  organice better game,js
