+++ app.js
@@ -1,8 +1,8 @@
-import menu from "./views/menu.js";
-import game from './views/game.js';
+
+import {game_init} from "./service/gameService.js";
 import qu from "./questions.js";
 //menu()
 //document.body.innerHTML="hh"
 qu();
 
-game();
+game_init();
