+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 
 <head>
-    <meta  charset="UTF-8" />
+    <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Reto JS - Q&A Game</title>
@@ -12,39 +12,23 @@
 </head>
 
 <body>
-    <div class="container-fluid d-flex h-100 flex-column">
+    <div>
         <header>
-            <div id="header" class="row p-2 header">
-                <div class="col">
-                    <h1>Preguntas y respuestas</h1>
-                </div>
+            <div id="header" class="header">
+                <h1>Preguntas y respuestas</h1>
             </div>
         </header>
 
-        <body>
-            <div class="container">
-                <div id="top" class="row m-5 justify-content-center"></div>
-            </div>
-            <div class="container">
-                <div id="middle" class="row m-5 justify-content-center"></div>
-            </div>
-            <div class="container">
-                <div id="bottom" class="row m-5 justify-content-center"></div>
-            </div>
-
-        </body>
-
-        <div></div>
+        <div id="body-div">
+        </div>
 
         <footer>
-            <div id="footer" class="row p-2 footer">
-                <div class="col">
-                    <h1>SofkaU</h1>
-                </div>
+            <div id="footer" class="footer">
+                <h1>SofkaU</h1>
             </div>
         </footer>
     </div>
-    <script type="module" src="./app.js"></script>
+    <script type="module" src="app.js"></script>
 </body>
 
 </html>
