+++ model/playerQuery.js
@@ -1,0 +1,5 @@
+
+  let player = {
+    name: "luis",
+    score: 0,
+  };
