+++ model/questionQuery.js
@@ -1,9 +1,9 @@
 //Get an object with all the question information
 async function getQuestionObject(index) {
     let questions = await JSON.parse(localStorage.getItem("Questions"));
-    let question = await questions[index];
+    let questionObject = await questions[index];
   
-    return question;
+    return questionObject;
   }
 
 // get a posible answer for a question
@@ -19,7 +19,7 @@ async function getChoice(index_question, index_choice) {
 async function getQuestion(index_round, index_question) {
     let questionObject = await getQuestionObject(index_question);
     let question = await questionObject.question;
-    return question;
+    return  question;
   }
   
 export {getQuestionObject, getChoice, getQuestion }
