+++ styles.css
@@ -1,16 +1,97 @@
+html,
 body {
-  background-color: #ffffff;
-  margin-bottom: 8%;
-  min-height: 100%;
+  background-color: #212529;
 }
 
 .header {
+  position: fixed;
   background-color: #e3a125;
+  width: 100%;
+  padding: 10px;
+  top: 0;
 }
 
 .footer {
   position: fixed;
-  width: 100%;
   background-color: #29a2ed;
-  bottom: 0;
+  width: 100%;
+  padding: 10px;
+  bottom: 0px;
+}
+
+table,
+tr,
+th {
+  width: 40%;
+  border: 1px;
+  margin: 10px;
+}
+
+table {
+  margin-right: auto;
+  margin-left: auto;
+  border: 1px solid;
+}
+
+th {
+  text-align: center;
+  border: 1px solid;
+  background-color: #1d53af;
+  padding: 10px;
+}
+
+td {
+  border-left: 1px solid;
+  text-align: center;
+  padding: 10px;
+}
+
+tr:nth-child(even) {
+  background-color: #9dbbda;
+}
+
+tr:nth-child(odd) {
+  background-color: #c9daeb;
+}
+
+tr:hover {
+  background-color: #e3a125;
+  font-weight: bold;
+}
+
+.buttons {
+  display: grid;
+  background-color: #1d53af;
+  border-radius: 5px;
+  border: 0;
+  margin-left: auto;
+  margin-right: auto;
+  margin-top: 40px;
+  margin-bottom: 40px;
+  width: 40%;
+  height: 40px;
+  font-weight: bold;
+}
+
+.buttons:hover {
+  background-color: #29a2ed;
+  color: #212529;
+  width: 41%;
+}
+
+.title {
+  color: #9dbbda;
+  display: flex;
+  justify-content: center;
+  margin-top: 20px;
+  margin-bottom: 20px;
+}
+
+#body-div {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: space-between;
+  margin: 100px;
+  height: 100%;
 }
